requirements.txt
crewai
crewai-tools 
load_dotenv
langchain-huggingface
# youtube-transcript-api
yt_dlp==2023.11.14 
youtube-transcript-api==0.6.1